refactored art page
